requirements.txt
streamlit
google-cloud-aiplatform
pandas
